+++ src/main/java/br/pucrio/les/esg_token_backend/services/value/IValueService.java
@@ -1,5 +1,11 @@
 package br.pucrio.les.esg_token_backend.services.value;
 
+import java.util.List;
+
+import br.pucrio.les.esg_token_backend.models.Value;
+
 public interface IValueService {
+
+    public List<Value> index();
 
 }