+++ src/main/java/br/pucrio/les/esg_token_backend/services/value/ValueServiceImpl.java
@@ -1,8 +1,11 @@
 package br.pucrio.les.esg_token_backend.services.value;
 
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import br.pucrio.les.esg_token_backend.models.Value;
 import br.pucrio.les.esg_token_backend.repository.IValueRepository;
 
 @Service
@@ -13,5 +16,10 @@ public class ValueServiceImpl implements IValueService {
 
     public ValueServiceImpl(IValueRepository valueRepository) {
         this.valueRepository = valueRepository;
+    }
+
+    @Override
+    public List<Value> index() {
+        return this.valueRepository.findAll();
     }
 }