+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Value JPA Repository
 - base Value Service
 - autowire Value Service with Value Repository integration
+- find all value service